# api_spec.yml
openapi: 3.0.3
info:
  title: 工廠訂單管理排程系統 API (Factory Order Management API)
  description: 提供訂單管理、排程調整、權限控制及視覺化功能之 API 介面規格。對應於專案的 Functional Requirements。
  version: 1.0.0
servers:
  - url: /api
    description: 主要 API 伺服器（Next.js App Router）
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      summary: 取得使用者清單
      description: |
        列出與目前 SUPERADMIN 相同 production type 的所有使用者。
        可選填 `role` query parameter 篩選角色。
      tags:
        - Permissions
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UserRole"
          description: 依角色篩選（SUPERADMIN / ADMIN / SALES）
      responses:
        "200":
          description: 使用者清單
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      summary: 邀請使用者
      description: |
        SUPERADMIN 建立 pending user 並寄出 180 秒有效的設定密碼邀請信。
      tags:
        - Permissions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "201":
          description: 邀請建立成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRole:
      type: string
      enum: [SUPERADMIN, ADMIN, SALES]
...
    User:
      type: object
      properties:
        id:
          type: string
        accountId:
          type: string
          nullable: true
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        group:
          type: string
...
    CreateUserInput:
      type: object
      required: [email, name, role, group]
      properties:
        email:
          type: string
          format: email
          example: admin-a4@mail.shangjung.com
        name:
          type: string
          example: Alice Wang
        role:
          $ref: "#/components/schemas/UserRole"
        group:
          type: string
          description: Production type (A / B / C)
          example: A
    InvitationCreateResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        group:
          type: string
        invitationExpiresAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: FORBIDDEN
        message:
          type: string
          example: You do not have permission to perform this action.
        details:
          type: object
          additionalProperties: true
  responses:
    Unauthorized:
      description: 未提供 token 或 token 無效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: UNAUTHORIZED
            message: Missing or invalid token
    Forbidden:
      description: 權限不足（role 或 scope 不符）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: FORBIDDEN
            message: "Role 'SALES' is not allowed. Required: SUPERADMIN."
...
    BadRequest:
      description: 請求格式錯誤或缺少必填欄位
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: BAD_REQUEST
            message: Invalid request body.
...